Replace PureComponent with Component handling its own open/closed state, and move Delete button into it
