+++ thoth-app/src/component/mod.rs
@@ -312,6 +312,7 @@ pub mod contributions_form;
 pub mod contributor;
 pub mod contributors;
 pub mod dashboard;
+pub mod delete_dialogue;
 pub mod funder;
 pub mod funders;
 pub mod fundings_form;
